Libraries Update

+++ Sources/OpenVPN3/openvpn/common/persistfile.hpp
@@ -4,7 +4,7 @@
 //               packet encryption, packet authentication, and
 //               packet compression.
 //
-//    Copyright (C) 2012-2020 OpenVPN Inc.
+//    Copyright (C) 2012-2022 OpenVPN Inc.
 //
 //    This program is free software: you can redistribute it and/or modify
 //    it under the terms of the GNU Affero General Public License Version 3
@@ -62,7 +62,7 @@ namespace openvpn {
       const ssize_t len = write_retry(fd(), buf, count);
       if (len < 0)
 	syserr("write");
-      if (len != count || len != ::lseek(fd(), 0, SEEK_CUR))
+      if (static_cast<size_t>(len) != count || len != ::lseek(fd(), 0, SEEK_CUR))
 	err("incomplete write");
       if (::ftruncate(fd(), len) < 0)
 	syserr("truncate");
